refactor: APIService 개편 및 로그아웃 로직 개편
